refactor: RandomNumberUtil 패키지 이동 및 Util 실패 테스트 수정
